% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-number-of-records.R
\name{get_number_of_records}
\alias{get_number_of_records}
\title{Retrieve the number of records that match a given IATI Datastore query}
\usage{
get_number_of_records(query_terms, collection = "activity")
}
\arguments{
\item{query_terms}{a \code{character} of length 1. The terms to search for.
Visit the API documentation at:
\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/how-to-use-the-datastore-api/}.}

\item{collection}{a \code{character} of length 1. One among \code{activity}, \code{budget},
\code{transaction}.}
}
\value{
The total number of records (\code{integer} of length 1) that match the
query.
}
\description{
This function sends a query to the IATI Datastore API
(\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/})
and returns the total number of records that match this query.

To learn how to write a query, users can read the documentation
available at:
\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/how-to-use-the-datastore-api/}.
}
\examples{
\dontrun{
## Search NGO in projects description ----
query <- "description_narrative:NGO"
get_number_of_records(query, collection = "activity")
}
}
