% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-records.R
\name{get_records}
\alias{get_records}
\title{Retrieve records that match a given IATI Datastore query}
\usage{
get_records(query_terms, collection = "activity", sleep = 1)
}
\arguments{
\item{query_terms}{a \code{character} of length 1. The terms to search for.
Visit the API documentation at:
\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/how-to-use-the-datastore-api/}.}

\item{collection}{a \code{character} of length 1. One among \code{activity}, \code{budget},
\code{transaction}.}

\item{sleep}{a \code{numeric} of length 1. The time interval between two
consecutive requests to not stress the API.}
}
\value{
A \code{data.frame} with records in rows and metadata in columns.
}
\description{
This function sends a query to the IATI Datastore API
(\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/})
and returns all records that match this query.

To learn how to write a query, users can read the documentation
available at:
\url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/how-to-use-the-datastore-api/}.
}
\examples{
\dontrun{
## Get records for NGO in projects description ----
query <- "description_narrative:NGO"
get_records(query, collection = "activity")
}
}
