% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riata-package.R
\docType{package}
\name{riati-package}
\alias{riati}
\alias{riati-package}
\title{riati: An R Client for the IATI Datastore API}
\description{
An interface for the IATI Datastore API \url{https://iatistandard.org/en/iati-tools-and-resources/iati-datastore/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/riati}
  \item Report bugs at \url{https://github.com/frbcesab/riati/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Stephanie D'Agata \email{stephanie.dagata@gmail.com} (\href{https://orcid.org/0000-0001-6941-8489}{ORCID})
}

}
\keyword{internal}
